feat: Add wifiManager class for managing WiFi connection
This commit adds a new class called wifiManager to handle WiFi connection management. The class includes methods for inputting the SSID and password, connecting to the WiFi network, and getting the connection status. This will be useful for configuring and managing WiFi settings in the application.
