+++ .vscode/c_cpp_properties.json
@@ -15,12 +15,12 @@
                 "/Users/kvnxpmac/Library/Arduino15/packages/esp32/tools/esp32-arduino-libs/idf-release_v5.1-442a798083/esp32/qio_qspi/include",
                 "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/cores/esp32",
                 "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/variants/esp32da",
+                "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/libraries/WiFi/src",
+                "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/libraries/Network/src",
                 "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/libraries/BluetoothSerial/src",
                 "/Users/kvnxpmac/Documents/Arduino/libraries/ModbusMaster/src",
                 "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/libraries/SPIFFS/src",
                 "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/libraries/FS/src",
-                "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/libraries/WiFi/src",
-                "/Users/kvnxpmac/Library/Arduino15/packages/esp32/hardware/esp32/3.0.1/libraries/Network/src",
                 "/Users/kvnxpmac/Documents/Arduino/libraries/ArduinoJson/src",
                 "/Users/kvnxpmac/Library/Arduino15/packages/esp32/tools/esp-x32/2302/xtensa-esp32-elf/include/c++/12.2.0",
                 "/Users/kvnxpmac/Library/Arduino15/packages/esp32/tools/esp-x32/2302/xtensa-esp32-elf/include/c++/12.2.0/xtensa-esp32-elf",
