+++ .vscode/tasks.json
@@ -1,33 +1,12 @@
 {
     "tasks": [
         {
-            "type": "cppbuild",
-            "label": "C/C++: clang build active file",
-            "command": "/usr/bin/clang",
-            "args": [
-                "-fcolor-diagnostics",
-                "-fansi-escape-codes",
-                "-g",
-                "${workspaceRoot}/src/*.cpp",
-                "-o",
-                "${workspaceRoot}/src/main"
-            ],
-            "options": {
-                "cwd": "${workspaceRoot}"
-            },
-            "problemMatcher": [
-                "$gcc"
-            ],
-            "group": "build",
-            "detail": "Task generated by Debugger."
-        },
-        {
             "label": "closeSerial",
             "command":"${command:arduino.closeSerialMonitor}",
         },
         {
             "label": "openSerial",
-            "command":"${command:arduino.openSerialMonitor}"
+            "command": "${command:arduino.openSerialMonitor}"
         }
     ],
     "version": "2.0.0"
