+++ src/ioManager.hpp
@@ -2,6 +2,7 @@
 #include "BluetoothSerial.h"
 #include "HardwareSerial.h"
 #include <WString.h>
+#pragma once
 
 class ioManager
 {
@@ -13,4 +14,6 @@ public:
     static String waitForInput();
     static int waitNumberInput();
     static BluetoothSerial *getSerialBT();
+    static void writeTextFile(const char *path, const char *text);
+    static String readTextFile(const char *path);
 };
