+++ src/wifi/wifiManager.hpp
@@ -1,0 +1,13 @@
+#include "WiFi.h"
+#include "../ioManager.hpp"
+#define printl ioManager::println
+
+class wifiManager
+{
+private:
+public:
+    void inputSSID();
+    void inputPassword();
+    void connect();
+    void getStatus();
+};
